+++ src/main/java/team/creative/littletiles/common/config/LittleTilesConfig.java
@@ -247,7 +247,7 @@ public class LittleTilesConfig {
         public void configured(Side side) {
             LittleGrid.loadGrid(base, scale, exponent, LittleTiles.CONFIG.general.defaultSelectedGrid);
             ItemMultiTiles.currentContext = LittleGrid.defaultGrid();
-            ItemLittleBag.maxStackSizeOfTiles = ItemLittleBag.maxStackSize * LittleGrid.overallDefault().count3d;
+            ItemLittleBag.maxStackSizeOfTiles = (int) (ItemLittleBag.maxStackSize * LittleGrid.overallDefault().count3d);
             LittlePremadeRegistry.reload();
             ItemMultiTiles.reloadExampleStructures();
         }
